requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
surya-ocr==0.8.0
torch==2.5.1
